+++ rio_stac/stac.py
@@ -141,8 +141,16 @@ def get_raster_info(
             "scale": src_dst.scales[band - 1],
             "offset": src_dst.offsets[band - 1],
         }
+
+        # If the Nodata is not set we don't forward it.
+        # If the Nodata is not finite (inf, NaN) we set it to None as mentioned
+        # in ECMA-262 (section 24.5.2, JSON.stringify, NOTE 4, page 683 of the ECMA-262 pdf at last edit):
+        # > Finite numbers are stringified as if by calling ToString(number).
+        # > NaN and Infinity regardless of sign are represented as the String null.
         if src_dst.nodata is not None:
-            value["nodata"] = src_dst.nodata
+            value["nodata"] = (
+                None if not numpy.isfinite(src_dst.nodata) else src_dst.nodata
+            )
 
         if src_dst.units[band - 1] is not None:
             value["unit"] = src_dst.units[band - 1]
@@ -256,7 +264,6 @@ def create_stac_item(
                 {
                     f"proj:{name}": value
                     for name, value in get_projection_info(src_dst).items()
-                    if value is not None
                 }
             )
             extensions.append(